digit-recognition/requirements.txt
fastapi
uvicorn
keras
tensorflow
scikit-learn
numpy
pillow
scikit-image
python-multipart
